Isolate from system letsencrypt by working in a dedicated directory
In order to keep certificates and private keys more secure (ref. #5),
dokku-letsencrypt will now create a private letsencrypt working
directory in /home/dokku/letsencrypt with minimal read permissions added
to allow the dokku user to access the contained certificate. If updating
from a previous version of the plugin, please make sure that the
`install` procedure is re-run and the letsencrypt working directory is
created.