Switch to using simp_le (fix #17, fix #13, fix #14, fix #15, closes #2)
This will make sure that the plugin will never request a TTY, making it
safe for use as part of a cronjob. An additional benefit is that simp_le
detects whether it is necessary to renew the certificate before
requesting a new one.

Important: Before dokku letsencrypt <app> can be run, the e-mail address
must now be specified using dokku letsencrypt:email <app> <e-mail>