Store certificates and keys in configuration-dependent directory
In cases where a previous let's encrypt certificate / account key is
available but the app URLs, let's encrypt account e-mail or the selected
server has changed, there might be problems with stale configuration
being used from the existing files to generate the new certificate.

By creating a SHA1 hash of all important arguments and using it to
determine a storage directory for certificates and keys, no stale
information will be re-used while it will be possible to renew a
certificate from a previous configuration, if so desired.