[Readme.md] Removed $ character in code examples
It's not copy-able-friendly when we use `$` in the code example, cause it makes people to actually copy the code and then paste in the console (if CRLF at the end it will automatically run and fail cause `$: command not found`). I find it easier and also understandable to just use the code without the `$` in the code. 
What do you think?